Move IGameFeatures to a separate file.
